+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Lint fix the generated files
+
 ## [0.4.3] - 2018-08-15
 
 ### Fixed